Finish question testing

+++ controllers/question.js
@@ -55,7 +55,7 @@ class QuestionController {
             if (question) {
                 res.status(200).json(question);
             } 
-            else{
+            else {
                 let err = { status: 404, message: `Question not found` };
                 next(err);
             }
@@ -96,7 +96,7 @@ class QuestionController {
             if (question) {
                 res.status(200).json(question);
             } 
-            else{
+            else {
                 let err = { status: 404, message: `Question not found` };
                 next(err);
             }
@@ -108,8 +108,14 @@ class QuestionController {
     
     static delete (req, res, next) {
         Question.findByIdAndDelete(req.params.id)
-        .then((question) => {                
-            return Answer.deleteMany({ QuestionId: req.params.id });
+        .then((question) => {            
+            if (question.deletedCount > 0) {
+                return Answer.deleteMany({ QuestionId: req.params.id });
+            }
+            else {
+                let err = { status: 404, message: `Question not found` };
+                next(err);
+            }
         })
         .then((answer) => {
             res.status(200).json({ message: "Question deleted successfully" });
@@ -122,13 +128,19 @@ class QuestionController {
     static solution(req, res, next) {
         Question.findById(req.params.id)
         .then((question) => {
-            if (question.solution != req.body.AnswerId) {
-                question.solution = req.body.AnswerId;
-                return question.save();
-            } 
+            if (question) {
+                if (question.solution != req.body.AnswerId) {
+                    question.solution = req.body.AnswerId;
+                    return question.save();
+                } 
+                else {
+                    question.solution = null;
+                    return question.save();
+                }
+            }
             else {
-                question.solution = null;
-                return question.save();
+                let err = { status: 404, message: `Question not found` };
+                next(err);
             }
         })
         .then((updated) => {
@@ -142,9 +154,15 @@ class QuestionController {
     static view(req, res, next) {
         Question.findById(req.params.id)
         .then((question) => {
-            if (question.views.includes(req.user._id) === false) {
-                question.views.push(req.user._id);
-                return question.save();
+            if (question) {
+                if (question.views.includes(req.user._id) === false) {
+                    question.views.push(req.user._id);
+                    return question.save();
+                }
+            }
+            else {
+                let err = { status: 404, message: `Question not found` };
+                next(err);
             }
         })
         .then((viewed) => {
@@ -158,22 +176,28 @@ class QuestionController {
     static upvote(req, res, next) {
         Question.findById(req.params.id)
         .then((question) => {
-            if (question.UserId != req.user._id) {
-                if (question.upvotes.includes(req.user._id) === false) {
-                    const downvote = question.downvotes.indexOf(req.user._id);
-                    if (downvote >= 0) {
-                        question.downvotes.splice(downvote, 1);
+            if (question) {
+                if (question.UserId != req.user._id) {
+                    if (question.upvotes.includes(req.user._id) === false) {
+                        const downvote = question.downvotes.indexOf(req.user._id);
+                        if (downvote >= 0) {
+                            question.downvotes.splice(downvote, 1);
+                        }
+                        question.upvotes.push(req.user._id);
+                        return question.save();
+                    } else {
+                        const index = question.upvotes.indexOf(req.user._id);
+                        question.upvotes.splice(index, 1);
+                        return question.save();
                     }
-                    question.upvotes.push(req.user._id);
-                    return question.save();
-                } else {
-                    const index = question.upvotes.indexOf(req.user._id);
-                    question.upvotes.splice(index, 1);
-                    return question.save();
+                }
+                else {
+                    let err = { status: 403, message: `You cannot vote your own question` };
+                    next(err);
                 }
             }
             else {
-                let err = { status: 400, message: `You cannot vote your own question` };
+                let err = { status: 404, message: `Question not found` };
                 next(err);
             }
         })
@@ -188,22 +212,28 @@ class QuestionController {
     static downvote(req, res, next) {
         Question.findById(req.params.id)
         .then((question) => {
-            if (question.UserId != req.user._id) {
-                if (question.downvotes.includes(req.user._id) === false) {
-                    const upvote = question.upvotes.indexOf(req.user._id);
-                    if (upvote >= 0) {
-                        question.upvotes.splice(upvote, 1);
+            if (question) {
+                if (question.UserId != req.user._id) {
+                    if (question.downvotes.includes(req.user._id) === false) {
+                        const upvote = question.upvotes.indexOf(req.user._id);
+                        if (upvote >= 0) {
+                            question.upvotes.splice(upvote, 1);
+                        }
+                        question.downvotes.push(req.user._id);
+                        return question.save();
+                    } else {
+                        const index = question.downvotes.indexOf(req.user._id);
+                        question.downvotes.splice(index, 1);
+                        return question.save();
                     }
-                    question.downvotes.push(req.user._id);
-                    return question.save();
-                } else {
-                    const index = question.downvotes.indexOf(req.user._id);
-                    question.downvotes.splice(index, 1);
-                    return question.save();
+                }
+                else {
+                    let err = { status: 403, message: `You cannot vote your own question` };
+                    next(err);
                 }
             }
             else {
-                let err = { status: 400, message: `You cannot vote your own question` };
+                let err = { status: 404, message: `Question not found` };
                 next(err);
             }
         })
